A new 'method' called 'method.stratifiedNNLS'
Proposing a new 'method' called 'method.stratifiedNNLS'. A single algorithm in the library (in the example: "SL.stratify") is used to data-adaptively identify subgroups in the observed data. The other library algorithms are fit with the full training data, but the new method allows the weights to vary by subgroup. So this differs from the subsemble method in that an overall predictor is fit and the weights can vary by subgroup, whereas subsemble estimates subset specific predictors.